Fix deprecated auth-tool in asreview.entry_points folder
